+++ threedi_models_and_simulations/api_calls/threedi_calls.py
@@ -59,6 +59,7 @@ from threedi_api_client.openapi import (
     SimulationStatus,
     SqliteFileUpload,
     StableThresholdSavedState,
+    Substance,
     TableStructureControl,
     Template,
     ThreediModel,
@@ -530,6 +531,11 @@ class ThreediCalls:
         breach = self.threedi_api.simulations_events_breaches_create(str(simulation_pk), data)
         return breach
 
+    def create_simulation_substances(self, simulation_pk: int, **data) -> Substance:
+        """Add substance to the given simulation."""
+        substance = self.threedi_api.simulations_substances_create(str(simulation_pk), data)
+        return substance
+
     def create_simulation_lateral_constant(self, simulation_pk: int, **data) -> ConstantLateral:
         """Add lateral constant to the given simulation."""
         lateral_constant = self.threedi_api.simulations_events_lateral_constant_create(str(simulation_pk), data)
@@ -925,9 +931,13 @@ class ThreediCalls:
         threedi_models = self.threedi_api.schematisations_revisions_threedimodels(revision_pk, schematisation_pk)
         return threedi_models
 
-    def create_schematisation_revision_3di_model(self, schematisation_pk: int, revision_pk: int) -> ThreediModel:
+    def create_schematisation_revision_3di_model(self, schematisation_pk: int, revision_pk: int, inherit_templates: bool = False) -> ThreediModel:
         """Create a new 3Di model out of committed revision."""
-        threedi_model = self.threedi_api.schematisations_revisions_create_threedimodel(revision_pk, schematisation_pk)
+        data = {
+            "inherit_from_previous_threedimodel": True,
+            "inherit_from_previous_revision": inherit_templates,
+        }
+        threedi_model = self.threedi_api.schematisations_revisions_create_threedimodel(revision_pk, schematisation_pk, data)
         return threedi_model
 
     def fetch_schematisation_revision_tasks(self, schematisation_pk: int, revision_pk: int) -> List[RevisionTask]: